+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 1.0.1
 =====
 
-
+- Fixed: `buildr clean package` would fail with a "do not know how to
+  build task [...]/target/gem_home" message.
 
 1.0.0
 =====